requirements.txt
fastapi
uvicorn
jinja2
python-multipart
tensorflow
pillow
numpy
